Fixed some issues.

+++ src/Timpex.Reporting.Filesetup/Timpex.Reporting.Filesetup/ByCore/IO/DirectoryImpl.cs
@@ -16,84 +16,96 @@ namespace Timpex.Reporting.Filesetup.ByCore.IO
         public virtual string CreateRootDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/";
+            CoreDirectory(path);
+            return path;
+        }
+
+        private void CoreDirectory(string path)
+        {
             Directory.CreateDirectory(path);
+        }
+
+        public virtual string Root()
+        {
+            var path = "C:/Devel/TRDeployment/";
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateSsisDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/SSIS/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateReportsDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/Reports/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateReportEnglishDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/Reports_ENG/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateReportEnglishUtvDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/Reports_ENG_UTV/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateReportNorUtvDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/Reports_UTV/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateDataSourceDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/DataSource/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateDataSetDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/DataSet/";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string  CreateSqlCmdDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/SQL/CMD";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateSqlRootDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/SQL";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateSqlMasterDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/SQL/Reporting.Master";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
         public virtual string CreateSqlUpdatesDirectory()
         {
             var path = "C:/Devel/TRDeployment/Reporting/SQL/Reporting.Updates";
-            Directory.CreateDirectory(path);
+            CoreDirectory(path);
             return path;
         }
 
@@ -104,7 +116,8 @@ namespace Timpex.Reporting.Filesetup.ByCore.IO
 
         public virtual void Remove(string path)
         {
-            Directory.Delete(path, true);
+            if(Directory.Exists(path))
+                Directory.Delete(path, true);
         }
     }
 }
